% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_pchn6_geo.R
\name{cbs_pchn6_geo}
\alias{cbs_pchn6_geo}
\title{Get the pchn6 tabel from cbs, GEO encoded with PDOK}
\usage{
cbs_pchn6_geo()
}
\value{
A dataframe.
}
\description{
This function gives back a dataframe containing the whole cbs pchn6 table enriched with the geolocation added by running pdok_find_coordinates on it. It has a 99\% + coverage
}
\examples{
df <-PDOK::cbs_pchn6_geo
}
