% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_query.R
\name{execute_query}
\alias{execute_query}
\title{Execute query
This function give you a dataframe which contains a query per dataste on PDOK.nl/datasets. Each function can be used within sf::st_read() to return a JSON containing vectors and data regarding the data.}
\usage{
execute_query(
  query_number,
  dataframe,
  query_column = "query_nr",
  link_column = "link",
  full
)
}
\arguments{
\item{use_stored_df}{A logical value. If TRUE, return the stored dataframe.
If FALSE, run a query.}
}
\value{
A dataframe.
}
\description{
Execute query
This function give you a dataframe which contains a query per dataste on PDOK.nl/datasets. Each function can be used within sf::st_read() to return a JSON containing vectors and data regarding the data.
}
\examples{
df <-PDOK::wfs_datasets(use_stored_df =TRUE)
df <-PDOK::wfs_datasets(use_stored_df =FALSE)
}
