% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_coordinates.R
\docType{data}
\name{input}
\alias{input}
\title{Get coordinates for Dutch addresses}
\format{
An object of class \code{character} of length 1.
}
\usage{
input
}
\arguments{
\item{input}{Input for coordinates, postcodes/ street names etc. Not all structures work, you should try a few to see what works for your inputdata.}

\item{verbose_succes}{If TRUE it will print the current input including the coordinates on a succesfull API call.}
}
\value{
a coordinates string
}
\description{
This function will query the PDOK API to get coordinates for your input. It first finds whether or not your input returns a suggestions. If so it will take the first one and find its coordinates. Using postcode: and huisnummer: gives good consistent results
}
\examples{
coordinates <-find_coordinates(input ="Amsterdam", verbose_succes = TRUE)
}
\keyword{datasets}
