% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdok_wfs_datasets.R
\name{pdok_wfs_datasets}
\alias{pdok_wfs_datasets}
\title{Get all WFS dataset descriptions and respective queries}
\usage{
pdok_wfs_datasets(stored_df = TRUE)
}
\arguments{
\item{use_stored_df}{A logical value. If TRUE, return the stored dataframe.
If FALSE, run a query.}
}
\value{
A dataframe.
}
\description{
This function gives back a dataframe with all WFS datasets from PDOK.nl and it's API calls. Usable with execute_query(). If it crashes on stored_DF = FALSE, it is most likely the service has a hickup.
}
\examples{
df <-PDOK::pdok_wfs_datasets(stored_df =TRUE)
df <-PDOK::pdok_wfs_datasets(stored_df =FALSE)
}
