% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdok_wfs_query.R
\name{pdok_wfs_query}
\alias{pdok_wfs_query}
\title{Execute query to get wfs geodata.}
\usage{
pdok_wfs_query(
  query_number,
  dataframe,
  query_column = "query_nr",
  link_column = "query",
  full = TRUE
)
}
\arguments{
\item{query_number}{The row related to the query in the query column}

\item{dataframe}{The dataframe from which to get your query, it can be called by using wfs_datasets()}

\item{query_column}{The column in which the queries are stored, defaults to 'query_nr'}

\item{link_column}{The column in which the queries are stored, defaults to 'query'}

\item{full}{TRUE/FALSE, if TRUE it will query until there are no more results left, if FALSE it will query max 1k once.}
}
\value{
A dataframe.
}
\description{
This function can execute queries that are stored in the dataframe that comes out of wfs_datasets(). It can do a single query or query the full underlying dataset.
}
\examples{
wfs_results <- pdok_wfs_query( query_number =1, dataframe = df, full=TRUE)
}
