% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbs_pchn6.R
\name{cbs_pchn6}
\alias{cbs_pchn6}
\title{Import Postcode/huisnummer tabels from cbs.nl}
\usage{
cbs_pchn6(jaar, add_names = TRUE)
}
\arguments{
\item{jaar}{A year value between from 2017 until 2024}

\item{add_names}{TRUE/FALSE. If TRUE it adds the names for buurt/wijk/gem instead of just the numeric identifier.}
}
\value{
A dataframe.
}
\description{
This function can download the custom postcode/huisnummer datasets from the corresponding webpages of cbs.nl. It downloads the zip to your working directory, unzips it and imports it into R. It then transforms it in the a structured DF.
}
\examples{
df <-PDOK::cbs_pchn6( jaar=2024, add_names= TRUE)
}
