% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_datasets.R
\name{wfs_datasets}
\alias{wfs_datasets}
\title{WFS Datasets}
\usage{
wfs_datasets(use_stored_df = FALSE)
}
\arguments{
\item{use_stored_df}{A logical value. If TRUE, return the stored dataframe.
If FALSE, run a query.}
}
\value{
A dataframe.
}
\description{
WFS Datasets
}
\examples{
PDOK::wfs_datasets(TRUE)
PDOK::wfs_datasets(FALSE)
}
