% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfs_datasets.R
\name{wfs_datasets}
\alias{wfs_datasets}
\title{WFS Datasets
This function give you a dataframe which contains a query per dataste on PDOK.nl/datasets. Each function can be used within sf::st_read() to return a JSON containing vectors and data regarding the data.}
\usage{
wfs_datasets(use_stored_df = FALSE)
}
\arguments{
\item{use_stored_df}{A logical value. If TRUE, return the stored dataframe.
If FALSE, run a query.}
}
\value{
A dataframe.
}
\description{
WFS Datasets
This function give you a dataframe which contains a query per dataste on PDOK.nl/datasets. Each function can be used within sf::st_read() to return a JSON containing vectors and data regarding the data.
}
\examples{
df <-PDOK::wfs_datasets(use_stored_df =TRUE)
df <-PDOK::wfs_datasets(use_stored_df =FALSE)
}
